+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## develop branch
 
-Nothing yet.
+### Fix
+
+* AllMatchingTypesList - when given an object, specialist types such as "String" and "Callable" now appear ahead of "Object" in the list
 
 ## 2.11.0 - Wed Sep 9 2015